implement role again 2

+++ app/Livewire/Activity/ActivityList.php
@@ -4,8 +4,10 @@ namespace App\Livewire\Activity;
 
 use App\Models\Activity;
 use App\Models\Company;
+use Illuminate\Support\Facades\Gate;
 use Livewire\Component;
 use Livewire\WithPagination;
+use Symfony\Component\HttpFoundation\Response;
 
 class ActivityList extends Component
 {
@@ -13,18 +15,31 @@ class ActivityList extends Component
     protected $paginationTheme = 'bootstrap';
     protected $listeners = ['refreshPage'];
     public $c, $q;
-    
+
     public function render()
     {
+        $permissions = [
+            'view-master-activity',
+            'create-master-activity',
+            'edit-master-activity',
+            'delete-master-activity',
+        ];
+        abort_if(Gate::none($permissions), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $companies = Company::all();
         $activitys = Activity::with(['company'])->search($this->q)
-        ->latest()
-        ->paginate(10);
-        return view('livewire.activity.activity-list', ['activitys' => $activitys,'companies' => $companies]);
+            ->latest()
+            ->paginate(10);
+        return view('livewire.activity.activity-list', ['activitys' => $activitys, 'companies' => $companies]);
     }
 
     public function delete($id)
     {
+        $permissions = [
+            'delete-master-activity',
+        ];
+        abort_if(Gate::none($permissions), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         try {
             Activity::where('id', $id)->delete();
             $this->dispatch('success', 'Data has been deleted');
